Cleanup windows post-install hack

+++ python/setup.py
@@ -4,15 +4,28 @@ import os.path
 # Available at setup time due to pyproject.toml
 from pybind11.setup_helpers import Pybind11Extension
 from setuptools import setup
+from setuptools.command.install import install
 
 __version__ = "0.0.1"
 
-# Handle improperly checked-out symlinks from git.
+# Handle improperly checked-out symlinks from git by manual copying
 # This happens for certain Windows users depending on git config
-if not os.path.isdir("src/cpp-core"):
+if not os.path.islink("src/cpp-core") and not os.path.isdir("src/cpp-core"):
     import shutil
+    contents = open("src/cpp-core", "rb").read()
     os.remove("src/cpp-core")
     shutil.copytree("../r/src/cpp-core", "src/cpp-core")
+    open("src/cpp-core/old-cpp-core", "wb").write(contents)
+
+class PostInstallHook(install):
+    def run(self):
+        install.run(self)
+        # Undo the manual copying of symlinks
+        if not os.path.islink("src/cpp-core") and os.path.isfile("src/cpp-core/old-cpp-core"):
+            import shutil
+            contents = open("src/cpp-core/old-cpp-core", "rb").read()
+            shutil.rmtree("src/cpp-core")
+            open("src/cpp-core", "wb").write(contents)
 
 
 ext_modules = [
@@ -29,6 +42,7 @@ ext_modules = [
 ]
 
 setup(
+    cmdclass={"install":PostInstallHook},
     ext_modules=ext_modules,
     extras_require={"test": "pytest"},
     version = __version__
